Publish API & Add Redeemer Cron
