+++ .dockerignore
@@ -1,9 +1,7 @@
-fly.*
 .git/
 .github/
 target
 *.log
 .env
-node_modules
 .vscode
 .idea
