+++ Cargo.toml
@@ -15,3 +15,4 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 dotenv = "0.15"
 chrono = "0.4"
 hex = "0.4.3"
+anyhow = "1.0.98"
