+++ src/main.rs
@@ -1,10 +1,13 @@
+mod api;
 mod db;
 mod models;
+mod redeem;
 
-use axum::{Json, Router, extract::State, routing::get};
+use axum::{Router, routing::get};
 use dotenv::dotenv;
 use sqlx::{PgPool, postgres::PgPoolOptions};
 use std::net::SocketAddr;
+use tokio::time::{Duration, interval};
 use tower_http::cors::{Any, CorsLayer};
 
 #[tokio::main]
@@ -25,15 +28,15 @@ async fn main() {
 
     // Build our application with a route
     let app = Router::new()
-        .route("/health", get(health_check))
-        .route("/redeemable", get(get_redeemable))
+        .route("/health", get(api::health_check))
+        .route("/redeemable", get(api::get_redeemable))
         .layer(
             CorsLayer::new()
                 .allow_origin(Any)
                 .allow_methods(Any)
                 .allow_headers(Any),
         )
-        .with_state(pool);
+        .with_state(pool.clone());
 
     // Get port from environment variable or default to 3000
     let port = std::env::var("PORT")
@@ -46,35 +49,16 @@ async fn main() {
     tracing::info!("listening on {}", addr);
 
     let listener = tokio::net::TcpListener::bind(addr).await.unwrap();
+    tokio::spawn(spawn_redeemer(pool.clone()));
     axum::serve(listener, app).await.unwrap();
 }
 
-async fn get_redeemable(State(pool): State<PgPool>) -> Json<Vec<models::RedeemableSubscription>> {
-    let current_timestamp = chrono::Utc::now().timestamp() as i32;
-    match db::get_redeemable_subscriptions(&pool, current_timestamp).await {
-        Ok(subscriptions) => Json(subscriptions),
-        Err(sqlx::Error::Database(db_err))
-            if db_err.code() == Some(std::borrow::Cow::Borrowed("42P01")) =>
-        {
-            // Table doesn't exist yet, return empty list
-            tracing::warn!("Database tables don't exist yet, returning empty list");
-            Json(Vec::new())
-        }
-        Err(e) => {
-            // Log other database errors but don't panic
-            tracing::error!("Database error: {}", e);
-            Json(Vec::new())
+async fn spawn_redeemer(pool: PgPool) {
+    let mut ticker = interval(Duration::from_secs(60 * 5)); // every 5 minutes
+    loop {
+        ticker.tick().await;
+        if let Err(err) = redeem::run_redeem_job(&pool).await {
+            tracing::error!("Redeem job failed: {err:?}");
         }
     }
-}
-
-async fn health_check(State(pool): State<PgPool>) -> Json<serde_json::Value> {
-    // Test database connectivity
-    let db_healthy = (sqlx::query("SELECT 1").execute(&pool).await).is_ok();
-
-    Json(serde_json::json!({
-        "status": if db_healthy { "healthy" } else { "unhealthy" },
-        "database": if db_healthy { "connected" } else { "disconnected" },
-        "timestamp": chrono::Utc::now().to_rfc3339()
-    }))
 }
