+++ src/redeem.rs
@@ -1,0 +1,6 @@
+use sqlx::PgPool;
+
+pub async fn run_redeem_job(_pool: &PgPool) -> anyhow::Result<()> {
+    // Insert core logic here from redeem-rs
+    Ok(())
+}
